+++ R/deployApp.R
@@ -301,10 +301,10 @@ deployApp <- function(appDir = getwd(),
     )
   }
   if (is.null(target$appId)) {
-    dest <- accountId(target$username, target$server)
+    dest <- accountLabel(target$username, target$server)
     taskComplete(quiet, "Deploying {.val {target$appName}} to {.val {dest}}")
   } else {
-    dest <- accountId(target$username, target$server)
+    dest <- accountLabel(target$username, target$server)
     taskComplete(quiet, "Re-deploying {.val {target$appName}} to {.val {dest}}")
   }
 
@@ -354,7 +354,7 @@ deployApp <- function(appDir = getwd(),
       }
     )
     if (application$id == target$appId) {
-      taskComplete(quiet, "Found application")
+      taskComplete(quiet, "Found application {.url {application$url}}")
     } else {
       taskComplete(quiet, "Created application with id {.val {application$id}}")
     }