+++ R/deployments-find.R
@@ -25,7 +25,7 @@ findDeployment <- function(appPath = ".",
 disambiguateDeployments <- function(appDeployments, error_call = caller_env()) {
   apps <- paste0(
     appDeployments$name, " ",
-    "(", accountId(appDeployments$account, appDeployments$server), "): ",
+    "(", accountLabel(appDeployments$account, appDeployments$server), "): ",
     "{.url ", appDeployments$url, "}"
   )
   not_interactive <- c(