+++ R/ide.R
@@ -3,13 +3,13 @@
 # This function is poorly named because as well as validating the server
 # url it will also register the server if needed.
 validateServerUrl <- function(url, certificate = NULL) {
-  valid <- validateConnectUrl(url, certificate)
+  res <- validateConnectUrl(url, certificate)
 
-  if (valid$ok)  {
+  if (res$valid)  {
     name <- findAndRegisterLocalServer(url)
-    c(list(valid = TRUE, url = valid$url, name = name), valid$response)
+    c(list(valid = TRUE, url = res$url, name = name), res$response)
   } else {
-    valid
+    res
   }
 }
 
@@ -50,12 +50,12 @@ validateConnectUrl <- function(url, certificate = NULL) {
     return(list(valid = FALSE, message = conditionMessage(cnd)))
   }
 
-  httpResponse <- attr(response, "httpResponse")
-  if (!isContentType(httpResponse, "application/json")) {
+  contentType <- attr(response, "httpContentType")
+  if (!isContentType(contentType, "application/json")) {
     return(list(valid = FALSE, message = "Endpoint did not return JSON"))
   }
 
-  url <- gsub("/server_settings$", "", buildHttpUrl(httpResponse$req))
+  url <- gsub("/server_settings$", "", attr(response, "httpUrl"))
   list(valid = TRUE, url = url, response = response)
 }