+++ R/utils.R
@@ -35,13 +35,6 @@ httpDiagnosticsEnabled <- function() {
     getOption("rsconnect.http.verbose", FALSE))
 }
 
-# wrapper around read.dcf to workaround LC_CTYPE bug
-readDcf <- function(...) {
-  loc <- Sys.getlocale("LC_CTYPE")
-  on.exit(Sys.setlocale("LC_CTYPE", loc))
-  read.dcf(...)
-}
-
 # Replacement for tools::file_path_sans_ext to work around an issue where
 # filenames like "foo..ext" are not returned as "foo.".
 file_path_sans_ext <- function(x, compression = FALSE) {